Moved the implementation
